Fix issue of last exercise of the day `sleep()` passed end time
The issue of having the last exercise of the day run *after* the specified end time is fixed by checking if the current time + the next random interval is still within the hours that are specified. This fixes #20.

Included with these fixes are a bunch of PEP8 changes; so style changes, but no _real_ code/logic changes.